[TASK] Move some tests for `Comment` to functional tests
Move the tests that rely on another class (`OutputFormat`)
(or might in the future).

Part of #757.
